React CSF types: Rename entry point, split out base types
