--- app/react/csf.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/client/preview/csf.d';
