+++ app/react/src/client/index.ts
@@ -10,6 +10,8 @@ export {
   forceReRender,
 } from './preview';
 
+export * from './preview/types-6-0';
+
 if (module && module.hot && module.hot.decline) {
   module.hot.decline();
 }
