+++ app/react/types-6-0.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/client/preview/types-6-0.d';
