+++ lib/cli/src/frameworks/react/ts/Button.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { Story, Meta } from '@storybook/react/csf';
+// also exported from '@storybook/react' if you can deal with breaking changes in 6.1
+import { Story, Meta } from '@storybook/react/types-6-0';
 
 import { Button, ButtonProps } from './Button';
